requirements.txt
#Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

#Data validation
pydantic>=2.7.0
pydantic-settings>=2.2.0

#LLM / LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.5
groq>=0.9.0

#Embeddings
huggingface-hub>=0.23.0
transformers>=4.41.0
sentence-transformers>=3.0.0

#Vector DB
chromadb>=0.5.0

#Search 
rank-bm25>=0.2.2

#Document loading
pypdf>=4.2.0          # PDF support
python-docx>=1.1.2    # DOCX support

#Utilities
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.2.0
python-dotenv>=1.0.0
loguru>=0.7.2
httpx>=0.27.0
requests>=2.32.0

#Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0         # FastAPI TestClient async support

#Evaluation
ragas>=0.1.0
datasets>=2.19.0

#mlflow
mlflow